requirements.txt
django
djangorestframework
requests
python-dotenv
psycopg2-binary 
django-cors-headers
Pillow